Fix SQL schema creation and refactor some code
In schema creation now the key is integer, reformat some spaces in code

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 create table if not exists  product (
-    id LONG NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    id INTEGER NOT NULL AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255),
     category VARCHAR(255),
     price DECIMAL(10,2),
